+++ tiledb/cc/common.cc
@@ -41,8 +41,8 @@ std::unordered_map<tiledb_datatype_t, std::string> _tdb_to_np_name_dtype = {
     {TILEDB_TIME_FS, "m8[fs]"},
     {TILEDB_TIME_AS, "m8[as]"},
 #endif
-#if TILEDB_VERSION_MAJOR >= 2 && TILEDB_VERSION_MINOR >= 7
-    {TILEDB_BLOB, "bytes"},
+#if TILEDB_VERSION_MAJOR >= 2 && TILEDB_VERSION_MINOR >= 9
+    {TILEDB_BLOB, "byte"},
 #endif
 #if TILEDB_VERSION_MAJOR >= 2 && TILEDB_VERSION_MINOR >= 10
     {TILEDB_BOOL, "bool"},
@@ -84,9 +84,6 @@ std::unordered_map<std::string, tiledb_datatype_t> _np_name_to_tdb_dtype = {
     {"timedelta64[ps]", TILEDB_TIME_PS},
     {"timedelta64[fs]", TILEDB_TIME_FS},
     {"timedelta64[as]", TILEDB_TIME_AS},
-#endif
-#if TILEDB_VERSION_MAJOR >= 2 && TILEDB_VERSION_MINOR >= 7
-    {"bytes", TILEDB_BLOB},
 #endif
 #if TILEDB_VERSION_MAJOR >= 2 && TILEDB_VERSION_MINOR >= 10
     {"bool", TILEDB_BOOL},